# artifacts/api-server/src/routes/alerts.ts
import { Router } from "express";
import { SubscribeAlertsBody } from "@workspace/api-zod";
import { db } from "@workspace/db";
import { alertSubscriptionsTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";

const router = Router();

router.post("/alerts/subscribe", async (req, res) => {
  const parsed = SubscribeAlertsBody.safeParse(req.body);
  if (!parsed.success) return res.status(400).json({ error: "Invalid subscription data" });
  const userId = req.user?.id ?? "anonymous";

  try {
    const [sub] = await db.insert(alertSubscriptionsTable).values({
      userId,
      endpoint: parsed.data.endpoint,
      p256dh: parsed.data.keys.p256dh,
      auth: parsed.data.keys.auth,
      minEdge: parsed.data.minEdge ?? 3,
      sports: parsed.data.sports ?? null,
      active: true,
    }).returning();

    return res.json({
      id: sub.id,
      status: "active",
      message: "Push alerts enabled. You will be notified when new value bets with your threshold are found.",
    });
  } catch (err) {
    req.log.error({ err }, "Error creating alert subscription");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;