# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdgeFinder sports betting value prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bets:
    get:
      operationId: getBets
      tags: [bets]
      summary: Get tracked bets for the current user
      parameters:
        - name: sport
          in: query
          schema:
            type: string
            enum: [NFL, NBA, MLB, NHL, all]
            default: all
        - name: result
          in: query
          schema:
            type: string
            enum: [win, loss, pending, all]
            default: all
      responses:
        "200":
          description: List of tracked bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedBet"
    post:
      operationId: createBet
      tags: [bets]
      summary: Log a new bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBetBody"
      responses:
        "201":
          description: Bet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedBet"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bets/{betId}:
    patch:
      operationId: updateBet
      tags: [bets]
      summary: Update a bet (e.g. set result after game)
      parameters:
        - name: betId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBetBody"
      responses:
        "200":
          description: Updated bet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedBet"
        "404":
          description: Bet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBet
      tags: [bets]
      summary: Delete a tracked bet
      parameters:
        - name: betId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Bet deleted
        "404":
          description: Bet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /alerts/subscribe:
    post:
      operationId: subscribeAlerts
      tags: [alerts]
      summary: Subscribe to push alerts for edge opportunities and odds shifts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertSubscribeBody"
      responses:
        "200":
          description: Subscription confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertSubscription"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        profileImage:
          type: string
        isAuthenticated:
          type: boolean
      required: [id, isAuthenticated]
...
    TrackedBet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        gameId:
          type: string
          nullable: true
        sport:
          type: string
          enum: [NFL, NBA, MLB, NHL]
        homeTeam:
          type: string
        awayTeam:
          type: string
        team:
          type: string
        betType:
          type: string
          enum: [moneyline, spread, over, under]
        bookmaker:
          type: string
        odds:
          type: integer
        stake:
          type: number
        result:
          type: string
          enum: [win, loss, pending]
        profit:
          type: number
          nullable: true
        gameDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, sport, homeTeam, awayTeam, team, betType, bookmaker, odds, stake, result, gameDate, createdAt]
    CreateBetBody:
      type: object
      properties:
        gameId:
          type: string
          nullable: true
        sport:
          type: string
          enum: [NFL, NBA, MLB, NHL]
        homeTeam:
          type: string
        awayTeam:
          type: string
        team:
          type: string
        betType:
          type: string
          enum: [moneyline, spread, over, under]
        bookmaker:
          type: string
        odds:
          type: integer
        stake:
          type: number
        gameDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required: [sport, homeTeam, awayTeam, team, betType, bookmaker, odds, stake, gameDate]
    UpdateBetBody:
      type: object
      properties:
        result:
          type: string
          enum: [win, loss, pending]
        notes:
          type: string
          nullable: true
...
    AlertSubscribeBody:
      type: object
      properties:
        endpoint:
          type: string
        keys:
          type: object
          properties:
            p256dh:
              type: string
            auth:
              type: string
          required: [p256dh, auth]
        minEdge:
          type: number
          default: 3
        sports:
          type: array
          items:
            type: string
      required: [endpoint, keys]
    AlertSubscription:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        message:
          type: string
      required: [id, status]