# lib/api-zod/src/generated/types/alertSubscribeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeFinder sports betting value prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSubscribeBodyKeys } from "./alertSubscribeBodyKeys";

export interface AlertSubscribeBody {
  endpoint: string;
  keys: AlertSubscribeBodyKeys;
  minEdge?: number;
  sports?: string[];
}