# lib/api-zod/src/generated/types/alertSubscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeFinder sports betting value prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface AlertSubscription {
  id: number;
  status: string;
  message?: string;
}