# lib/api-zod/src/generated/types/createBetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeFinder sports betting value prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBetBodyBetType } from "./createBetBodyBetType";
import type { CreateBetBodySport } from "./createBetBodySport";

export interface CreateBetBody {
  gameId?: string | null;
  sport: CreateBetBodySport;
  homeTeam: string;
  awayTeam: string;
  team: string;
  betType: CreateBetBodyBetType;
  bookmaker: string;
  odds: number;
  stake: number;
  gameDate: Date;
  notes?: string | null;
}